refactor: update import paths for ctaLinks in 2024 components
